+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - sqlmath - add sql-function fill_forward()
 
 # v2023.4.22
+- sqlmath - add sql-function avg_ema()
 - sqlmath - rename sql-function percentile() to quantile()
 - ci - add shell-function shCiPreCustom2() to setup python in custom-ci
 - ci - remove mandatory python ci